Rename maker::rollover to legacy_rollover
We're still temporarily supporting rollovers over the legacy network
stack in the maker. It's clearer if the name is more explicit.
