+++ maker/src/connection.rs
@@ -1,7 +1,7 @@
 use crate::cfd;
 use crate::collab_settlement;
 use crate::contract_setup;
-use crate::rollover;
+use crate::legacy_rollover;
 use anyhow::bail;
 use anyhow::Context;
 use anyhow::Result;
@@ -105,7 +105,7 @@ pub struct TakerMessage {
 
 pub struct RegisterRollover {
     pub order_id: OrderId,
-    pub address: xtra::Address<rollover::Actor>,
+    pub address: xtra::Address<legacy_rollover::Actor>,
 }
 
 pub struct Actor {
@@ -118,7 +118,7 @@ pub struct Actor {
     p2p_socket: SocketAddr,
     setup_actors: AddressMap<OrderId, contract_setup::Actor>,
     settlement_actors: AddressMap<OrderId, collab_settlement::Actor>,
-    rollover_actors: AddressMap<OrderId, rollover::Actor>,
+    rollover_actors: AddressMap<OrderId, legacy_rollover::Actor>,
 }
 
 /// A connection to a taker.
@@ -584,7 +584,7 @@ impl Actor {
             RolloverProtocol { order_id, msg } => {
                 if let Err(NotConnected(_)) = self
                     .rollover_actors
-                    .send_async(&order_id, rollover::ProtocolMsg(msg))
+                    .send_async(&order_id, legacy_rollover::ProtocolMsg(msg))
                     .await
                 {
                     tracing::warn!(%order_id, "No active rollover actor");
