+++ maker/src/lib.rs
@@ -14,8 +14,8 @@ pub mod cfd;
 mod collab_settlement;
 mod connection;
 mod contract_setup;
+mod legacy_rollover;
 mod metrics;
-mod rollover;
 pub mod routes;
 
 #[derive(Parser)]
